+++ src/hooks/useMouseDrag.ts
@@ -1,0 +1,57 @@
+import React, { useState } from 'react';
+
+function useMouseDrag(scrollRef: any) {
+  const [isDrag, setIsDrag] = useState(false);
+  const [startX, setStartX] = useState(0);
+
+  const onDragStart = (e: React.MouseEvent<HTMLDivElement>) => {
+    e.preventDefault();
+    if (scrollRef.current) {
+      setIsDrag(true);
+      setStartX(e.pageX + scrollRef.current?.scrollLeft);
+    }
+  };
+
+  const onDragEnd = () => {
+    setIsDrag(false);
+  };
+
+  const onDragMove = (e: React.MouseEvent<HTMLDivElement>) => {
+    if (isDrag) {
+      const { scrollWidth, clientWidth, scrollLeft } = scrollRef.current;
+
+      scrollRef.current.scrollLeft = startX - e.pageX;
+
+      if (scrollLeft === 0) {
+        setStartX(e.pageX);
+      } else if (scrollWidth <= clientWidth + scrollLeft) {
+        setStartX(e.pageX + scrollLeft);
+      }
+    }
+  };
+
+  const throttle = (func: any, ms: number) => {
+    let throttled = false;
+    return (...args: any[]) => {
+      if (!throttled) {
+        throttled = true;
+        setTimeout(() => {
+          func(...args);
+          throttled = false;
+        }, ms);
+      }
+    };
+  };
+
+  const delay = 100;
+  const onThrottleDragMove = throttle(onDragMove, delay);
+
+  return {
+    isDrag,
+    onDragStart,
+    onDragEnd,
+    onThrottleDragMove,
+  };
+}
+
+export default useMouseDrag;
